more ami

+++ java/src/IceInternal/Outgoing.java
@@ -216,9 +216,9 @@ public final class Outgoing
 	assert(_state <= StateInProgress);
 	
 	_is.swap(is);
-	byte status = _is.readByte();
+	int status = (int)_is.readByte();
 	
-	switch((int)status)
+	switch(status)
 	{
 	    case DispatchStatus._DispatchOK:
 	    {
